+++ example/server-drizzle/src/trpc/routers/__tests__/comment.test.ts
@@ -1,17 +1,20 @@
 import { beforeEach, expect, test, vi } from 'vite-plus/test';
 
-const { deleteCommentRecord, fetchCommentById, fetchPostById } = vi.hoisted(() => ({
+const { deleteCommentRecord, fetchById } = vi.hoisted(() => ({
   deleteCommentRecord: vi.fn(),
-  fetchCommentById: vi.fn(),
-  fetchPostById: vi.fn(),
+  fetchById: vi.fn(),
 }));
 
 vi.mock('../../../drizzle/queries.ts', () => ({
   createCommentRecord: vi.fn(),
   deleteCommentRecord,
-  fetchCommentById,
-  fetchPostById,
-  searchCommentsConnection: vi.fn(),
+}));
+
+vi.mock('../../executor.ts', () => ({
+  drizzleRegistry: new Map(),
+  drizzleRuntime: {
+    fetchById,
+  },
 }));
 import { router } from '../../init.ts';
 import { commentRouter } from '../comment.ts';
@@ -21,22 +24,23 @@ beforeEach(() => {
 });
 
 test('delete returns the post relation after the comment has been removed', async () => {
-  fetchCommentById.mockResolvedValue({
-    authorId: 'user-1',
-    content: 'hello',
-    id: 'comment-1',
-    post: {
-      _count: { comments: 3 },
+  fetchById
+    .mockResolvedValueOnce({
+      authorId: 'user-1',
+      content: 'hello',
+      id: 'comment-1',
+      post: {
+        _count: { comments: 3 },
+        id: 'post-1',
+        title: 'Post title',
+      },
+      postId: 'post-1',
+    })
+    .mockResolvedValueOnce({
+      _count: { comments: 2 },
       id: 'post-1',
       title: 'Post title',
-    },
-    postId: 'post-1',
-  });
-  fetchPostById.mockResolvedValue({
-    _count: { comments: 2 },
-    id: 'post-1',
-    title: 'Post title',
-  });
+    });
 
   const appRouter = router({ comment: commentRouter });
   const caller = appRouter.createCaller({
@@ -51,9 +55,13 @@ test('delete returns the post relation after the comment has been removed', asyn
   });
 
   expect(deleteCommentRecord).toHaveBeenCalledWith('comment-1');
-  expect(fetchPostById).toHaveBeenCalledWith('post-1', expect.objectContaining({}));
+  expect(fetchById).toHaveBeenCalledWith(
+    expect.objectContaining({
+      id: 'post-1',
+    }),
+  );
   expect(deleteCommentRecord.mock.invocationCallOrder[0]).toBeLessThan(
-    fetchPostById.mock.invocationCallOrder[0],
+    fetchById.mock.invocationCallOrder[1],
   );
   expect(result).toEqual({
     content: 'hello',
